fix: 处理自动化审核问题
